convert evaluateJob to a wrapper job
